Switch flags to use css flags
this gives us a lot more countries than having to manually map, download and
maintain all countries in the styl file. now we *only* map, download and
maintain languages that don't map elegantly to a country/flag (like arabic,
occitan, or catalan).

this is WIP as a lot of languages are not detected or get the wrong flag,
but i feel it's a very good path forward.

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>